Move Common code to separate file
